+++ aws-xray/src/test/java/io/opentelemetry/contrib/awsxray/AwsSpanMetricsProcessorTest.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.contrib.awsxray;
 
-import static io.opentelemetry.semconv.SemanticAttributes.HTTP_RESPONSE_STATUS_CODE;
+import static io.opentelemetry.semconv.trace.attributes.SemanticAttributes.HTTP_STATUS_CODE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -141,7 +141,7 @@ class AwsSpanMetricsProcessorTest {
 
   @Test
   public void testOnEndMetricsGenerationWithoutMetricAttributes() {
-    Attributes spanAttributes = Attributes.of(HTTP_RESPONSE_STATUS_CODE, 500L);
+    Attributes spanAttributes = Attributes.of(HTTP_STATUS_CODE, 500L);
     ReadableSpan readableSpanMock = buildReadableSpanMock(spanAttributes);
     Attributes metricAttributes = buildMetricAttributes(CONTAINS_NO_ATTRIBUTES);
     configureMocksForOnEnd(readableSpanMock, metricAttributes);
@@ -154,7 +154,7 @@ class AwsSpanMetricsProcessorTest {
 
   @Test
   public void testOnEndMetricsGenerationWithoutEndRequired() {
-    Attributes spanAttributes = Attributes.of(HTTP_RESPONSE_STATUS_CODE, 500L);
+    Attributes spanAttributes = Attributes.of(HTTP_STATUS_CODE, 500L);
     ReadableSpan readableSpanMock = buildReadableSpanMock(spanAttributes);
     Attributes metricAttributes = buildMetricAttributes(CONTAINS_ATTRIBUTES);
     configureMocksForOnEnd(readableSpanMock, metricAttributes);
@@ -167,7 +167,7 @@ class AwsSpanMetricsProcessorTest {
 
   @Test
   public void testOnEndMetricsGenerationWithLatency() {
-    Attributes spanAttributes = Attributes.of(HTTP_RESPONSE_STATUS_CODE, 200L);
+    Attributes spanAttributes = Attributes.of(HTTP_STATUS_CODE, 200L);
     ReadableSpan readableSpanMock = buildReadableSpanMock(spanAttributes);
     Attributes metricAttributes = buildMetricAttributes(CONTAINS_ATTRIBUTES);
     configureMocksForOnEnd(readableSpanMock, metricAttributes);
@@ -245,7 +245,7 @@ class AwsSpanMetricsProcessorTest {
 
   private static ReadableSpan buildReadableSpanWithThrowableMock(Throwable throwable) {
     // config http status code as null
-    Attributes spanAttributes = Attributes.of(HTTP_RESPONSE_STATUS_CODE, null);
+    Attributes spanAttributes = Attributes.of(HTTP_STATUS_CODE, null);
     ReadableSpan readableSpanMock = mock(ReadableSpan.class);
     SpanData mockSpanData = mock(SpanData.class);
     InstrumentationScopeInfo awsSdkScopeInfo =
@@ -280,7 +280,7 @@ class AwsSpanMetricsProcessorTest {
 
   private void validateMetricsGeneratedForHttpStatusCode(
       Long httpStatusCode, ExpectedStatusMetric expectedStatusMetric) {
-    Attributes spanAttributes = Attributes.of(HTTP_RESPONSE_STATUS_CODE, httpStatusCode);
+    Attributes spanAttributes = Attributes.of(HTTP_STATUS_CODE, httpStatusCode);
     ReadableSpan readableSpanMock = buildReadableSpanMock(spanAttributes);
     Attributes metricAttributes = buildMetricAttributes(CONTAINS_ATTRIBUTES);
     configureMocksForOnEnd(readableSpanMock, metricAttributes);